New: hidden-argument andEnable
`Clash.Explicit.Signal.enable` is renamed to `andEnable` and the
existing name is deprecated. A new `Clash.Signal.andEnable` is added
that performs the same function but using a `HiddenEnable` instead.

Documentation in Clash.Signal is corrected and extended. Parts of the
existing documentation were not rendered before this commit due to
incorrect interaction between CPP #ifdef and Haddock syntax.

The documentation has a new section called
_Monomorphism restriction leads to surprising behavior_
that warns users for issues with implicit parameters and offers advice
on how to avoid them.

A developer note [Going from WithSingleDomain to WithSpecificDomain] is
added in the source, explaining the code.