Expose pure variant of isWorkFree
When changing the RewriteState to use MVar, the isWorkFree function
causes something of a problem in it's current form, as it forms a
recursive group with isWorkFreeBinder. It cannot simply take a lens
to an MVar, as this is not the cache itself, but the means to
obtain the cache.

To make things simpler, the work free functions are rewritten to be
pure, and the old lensy monadic interface is exposed as another
function on top.